Feat/multiple sources (#45)
* adding ability to merge sources, json output

* removing comment

* adding rate support for sources

* remove comments

+++ models/config.go
@@ -84,7 +84,7 @@ type Field struct {
 	Function   string     `json:"function,omitempty"`
 	Source     SourceList `json:"source,omitempty"`
 	Template   string     `json:"template,omitempty"`
-	Rate       *int       `json:"rate,omitempty,string"`
+	Rate       *float64   `json:"rate,omitempty,string"`
 	Regex      string     `json:"regex,omitempty"`
 	Fields     []Field    `json:"fields,omitempty"`
 	Repeat     string     `json:"repeat,omitempty"`
